requirements.txt
torch>=2.1.0
numpy
pyyaml
scipy
statsmodels
tensorboard
